changed user service test cases
